69-backend-student-search-for-job-postings - Changed search parameters

+++ Backend/TheHiveApp/src/main/java/com/example/thehiveapp/entity/jobPosting/JobPosting.java
@@ -65,7 +65,7 @@ public class JobPosting {
     private JobType jobType;
 
     @Column(name = "minimum_gpa")
-    private BigDecimal minimumGpa;
+    private Double minimumGpa;
 
     @Column(name = "job_start")
     private LocalDate jobStart;
